Added lots of functionality
Currently can query and save multiple searches. These are saved as pickle objects in saved_searches/.

Currently in testing mode, does not query the website, uses output1.txt for HTML parsing

update() for listings class is currently not working, does not add multiple prices or times. Also adds new info even if price hasn't changed, this needs to be fixed.

Need to implement a way to check if a car still exists or has been sold or is unavailable. Could loop through objects not found in most recent search, which have not been marked as sold or unavailable in the script. This should do the trick.